intit other

+++ com/poojahshah/example/chaper/eigth/Answer.java
@@ -22,6 +22,6 @@ class Answer {
         // Your code goes here.
         int total = source.stream().mapToInt(Integer::intValue).sum();
         int totalNew = source.stream().reduce(0,Integer::sum);
-        return source.stream().mapToInt(Integer::intValue).sum() / source.size();
+        return source.stream().mapToInt(Integer::intValue).sum() / (double) source.size();
     }
 }
